refactor: ディレクトリ構成色々と見直し
