+++ src/app/lefmarna-otoiawase/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { LefmarnaOtoiawase } from './LefmarnaOtoiawase'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('お問い合わせ')
