+++ src/app/password/forgot/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { PasswordForgot } from './PasswordForgot'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('パスワード再設定の申請')
