+++ src/app/pokemons/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { PokemonIndex } from './PokemonIndex'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('みんなの投稿')
