+++ src/app/pokemons/show/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { PokemonShow } from './PokemonShow'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('ポケモン詳細')
