+++ src/app/privacy-policy/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { PrivacyPolicy } from './PrivacyPolicy'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('利用規約')
