+++ src/app/register/resend/page.tsx
@@ -1,5 +1,5 @@
-import { getMetadata } from '@/libs/metadata'
 import { RegisterResend } from './RegisterResend'
+import { getMetadata } from '@/utils/helpers'
 
 export const metadata = getMetadata('メール確認')
