+++ src/utils/helpers/index.ts
@@ -1,6 +1,11 @@
 export { requestApi } from './apiHelpers'
 export { exceptionErrorToArray } from './errorHelpers'
 export {
+  DEFAULT_META_TITLE,
+  DEFAULT_META_DESCRIPTION,
+  getMetadata,
+} from './metaHelpers'
+export {
   convertToInteger,
   numberToInt,
   valueVerification,
